Add external module
